+++ R/VWRfirstrun.R
@@ -4,9 +4,10 @@
 ############################################################################################################################
 #' @title VertexWiseR system requirements installation
 #'
-#' @description Helps the user verify if VertexWisrR's system requirements are present and install them ('Miniconda', 'BrainStat' toolbox and libraries). If they are installed already, nothing will be overwritten. 
+#' @description Helps the user verify if VertexWisrR's system requirements are present and install them (a suitable 'Python' or 'Miniconda' environment, 'BrainStat' toolbox and libraries). If they are installed already, nothing will be overwritten. 
 #'
-#' @details VertexWiseR imports and makes use of the R package 'reticulate.' 'reticulate' is a package that allows R to borrow or translate Python functions into R. Using 'reticulate', the package calls functions from the 'BrainStat' Python module. For 'reticulate' to work properly with VertexWiseR, 'Miniconda' needs to be installed with it — 'Miniconda' is a lightweight version of Python, specifically for use within 'RStudio'. If for a reason Miniconda cannot be installed, the function gives the choice to install a reticulate-suitable Python environment.
+#' @details VertexWiseR imports and makes use of the R package 'reticulate.' 'reticulate' is a package that allows R to borrow or translate Python functions into R. Using 'reticulate', the package calls functions from the 'BrainStat' Python module. For 'reticulate' to work properly with VertexWiseR, a Python environment needs to be installed with it — the default choice offered by VWRfirstrun is to let reticulate (version 1.41.0) create an ephemeral Python virtual environment using UV and py_require(). 
+#' If for a reason this is not desirable, VWRfirstrun() also gives the choice to install a 'Miniconda' (lightweight version of Python) or Python environment in a reticulate default path or a specified path.
 #' Vertex-wise statistical analyses of cortical surface require fsaverage and parcellation templates as imported by default in 'BrainStat'. 
 #' The decode_surf_data() function also requires the 'Neurosynth' database to be downloaded.
 #' @param requirement String that specifies a requirement to enquire about: 
@@ -20,7 +21,7 @@
 #' @return No returned value in interactive session. In non-interactive sessions, a string object informing that system requirements are missing.
 #' @examples
 #' VWRfirstrun()
-#' @importFrom reticulate conda_binary py_module_available miniconda_path
+#' @importFrom reticulate conda_binary py_module_available miniconda_path py_require
 #' @importFrom fs path_home
 #' @importFrom methods is 
 #' @importFrom utils menu
@@ -48,7 +49,17 @@ VWRfirstrun=function(requirement="any", n_vert=0, promptless=FALSE)
   # them from the package directory:
   Renvironpath=paste0(tools::R_user_dir(package='VertexWiseR'),'/.Renviron')
   if (file.exists(Renvironpath)) {
-    readRenviron(Renvironpath)}
+    readRenviron(Renvironpath)
+  
+  #if cache was cleared and UV environment no longer exists, remove it
+  if (!file.exists(Sys.getenv('VIRTUAL_ENV')))
+  {
+    Sys.unsetenv('VIRTUAL_ENV')
+    lines = readLines(Renvironpath) 
+    lines = lines[!grepl("^VIRTUAL_ENV", lines)]
+    writeLines(lines, Renvironpath)
+  }
+  }
   
   #default time limit to download is 60s which can be too short:
   options(timeout=500); #set to 500s instead
@@ -62,19 +73,45 @@ VWRfirstrun=function(requirement="any", n_vert=0, promptless=FALSE)
     #check if miniconda (or suitable python environment) is installed
     message('Checking for Miniconda or Python environment...')
     
-    if(is(tryCatch(reticulate::py_config(), error=function(e) e))[1] == 'simpleError') #fast, though less reliable check if Python is in a custom location. It will work if VWRfirstrun() was run once already
+    if(reticulate::py_available()==FALSE) #fast, but unreliable check if Python is in a custom location. It will work if VWRfirstrun() was run once already
     {
       
       if (is.null(tryCatch(reticulate::py_discover_config(), error = function(e) NULL))) #slow but reliable check if first time 
       {
         missingobj=1
         
-        prompt = utils::menu(c("Yes, install Miniconda (Recommended)",
-                               "Yes, install Python (if Miniconda could not be installed)", 
-                               "No"), 
-                             title="Miniconda or a suitable version of Python for reticulate could not be found in the environment. \n Do you want Miniconda or Python to be installed now?")
-        
-        if (prompt==1) #Install Miniconda
+        prompt = utils::menu(title="A suitable Python environment for reticulate or Miniconda could not be found in the environment. \n Do you want Miniconda or Python to be installed now?",
+                  choices = c("Yes, install a virtual Python environment (Recommended)",
+                              "Yes, install Miniconda",
+                              "Yes, install Python", 
+                              "No"), 
+                             )
+        if (prompt==1) #Install ephemeral virtual environment via UV 
+        { 
+          #check if virtual environment available with the right settings
+          #and initialize it.
+          message('Installing Python ephemeral environment via reticulate\'s py_require() and UV...\n')
+          reticulate::py_require(packages=c("numpy<=1.26.4","matplotlib","brainstat==0.4.2","vtk==9.3.1"), python_version = "<3.11")
+          reticulate::py_config()
+          
+          #will store cache path in .Renviron in tools::R_user_dir() 
+          #location specified by CRAN, create it if not existing:
+          envpath=tools::R_user_dir(package='VertexWiseR')
+          if (!dir.exists(envpath)) {dir.create(envpath, 
+                                                recursive = TRUE)} 
+          #make .Renviron file there and set conda/python paths:
+          renviron_path <- file.path(envpath, ".Renviron")
+          env_vars <- c(paste0('VIRTUAL_ENV="',
+                               Sys.getenv('VIRTUAL_ENV'),'"'))
+          cat(paste0(env_vars, "\n", collapse = "\n"), 
+              file = renviron_path, sep = "\n", append = TRUE)
+          
+          message(paste0('\nThis virtual environment is cached in the following path:\n', tools::R_user_dir("reticulate", "cache"), 
+                         '\n\nIf it gets cleared, VWRfirstrun() will prompt you again for a new installation. Unless the cache is removed, it will NOT use a new virtual environment created via py_require(), as the virtual path is fixated there:\n',
+                         paste0(tools::R_user_dir(package='VertexWiseR'),'/.Renviron')
+                         ,'\n'))
+        } 
+        else if (prompt==2) #Install Miniconda
         {
           
           #define default miniconda directory 
@@ -102,16 +139,26 @@ VWRfirstrun=function(requirement="any", n_vert=0, promptless=FALSE)
                                                   recursive = TRUE) }
             
             #get path to python executable (will differ across OS)
+            #silenced with sink and con
             Sys.setenv(RETICULATE_PYTHON=defaultpath)
             defaultpathexe <- tryCatch({
               tmpfile <- tempfile()
               con <- file(tmpfile, open = "wt")
               sink(con, type = "message")
-              on.exit({
-                sink(type = "message");
-                close(con); file.remove(tmpfile)}, add = TRUE)
-              reticulate::py_discover_config()$python
-            }, error = function(e) {NULL})
+              result <- reticulate::py_discover_config()$python  
+              # Close sink and connection 
+              sink(type = "message")
+              close(con)
+              file.remove(tmpfile)
+              result
+            }, error = function(e) {
+              # Ensure cleanup in case of error
+              sink(type = "message")
+              close(con)
+              file.remove(tmpfile)
+              NULL
+            })
+            
             
             #make .Renviron file there and set conda/python paths:
             renviron_path <- file.path(envpath, ".Renviron")
@@ -167,18 +214,28 @@ VWRfirstrun=function(requirement="any", n_vert=0, promptless=FALSE)
             Sys.setenv(RETICULATE_PYTHON_ENV=userpath)
             reticulate::py_install("vtk==9.3.1",pip = TRUE) # latest vtk==9.4.0 causes problems
             
-            #add python executable to the new installation 
-            #path to it will differ across OS:
-            Sys.setenv(RETICULATE_PYTHON=userpath)
+            #get path to python executable (will differ across OS)
+            #silenced with sink and con
+            Sys.setenv(RETICULATE_PYTHON=defaultpath)
             userpathexe <- tryCatch({
               tmpfile <- tempfile()
               con <- file(tmpfile, open = "wt")
               sink(con, type = "message")
-              on.exit({
-                sink(type = "message");
-                close(con); file.remove(tmpfile)}, add = TRUE)
-              reticulate::py_discover_config()$python
-            }, error = function(e) {NULL})
+              result <- reticulate::py_discover_config()$python  
+              # Close sink and connection 
+              sink(type = "message")
+              close(con)
+              file.remove(tmpfile)
+              result
+            }, error = function(e) {
+              # Ensure cleanup in case of error
+              sink(type = "message")
+              close(con)
+              file.remove(tmpfile)
+              NULL
+            })
+            
+            #add user path to the .Renviron:
             env_vars <- paste0('RETICULATE_PYTHON="',
                                userpathexe,'"')
             # Write to the .Renviron file and read again
@@ -191,12 +248,12 @@ VWRfirstrun=function(requirement="any", n_vert=0, promptless=FALSE)
         message('Please restart R after Miniconda installation for its environment to be properly detected by reticulate.')
           
         }
-        else if (prompt==2) #Install Python instead of Miniconda
+        else if (prompt==3) #Install Python instead of Miniconda
         { 
           
           #give the choices to specify path
-          choice = utils::menu(c("Default", "Custom"), 
-                               title=paste0("Reticulate's Python default installation is done within 'r-reticulate' in your OS libraries, via install_python(). Type \"1\" or \"Default\" if you want the default installation. \nYou can, alternatively, specify the path where Python will be installed."))
+          choice = utils::menu(title=paste0("Reticulate's Python default installation is done within 'r-reticulate' in your OS libraries, via install_python(). Type \"1\" or \"Default\" if you want the default installation. \nYou can, alternatively, specify the path where Python will be installed."),
+                               choices=c("Default", "Custom"))
           
           if (choice==1) #Install Python within default path
           {
@@ -212,21 +269,29 @@ VWRfirstrun=function(requirement="any", n_vert=0, promptless=FALSE)
           #Read new python enviroment
           Renvironpath=paste0(tools::R_user_dir(package='VertexWiseR'),'/.Renviron')
           if (file.exists(Renvironpath)) {readRenviron(Renvironpath)}
+          
+          #vtk installation
           message('A specific version of the vtk package (v9.3.1) will be installed.\n')
           #latest vtk==9.4.0 causes problems
-          status <- system("pip install vtk==9.3.1");
+          status <- system(paste0(Sys.getenv('RETICULATE_PYTHON')," -m pip install vtk==9.3.1"));
           #if failed then try pip3 
           if (status != 0) { 
             message('Could not install package with pip, trying pip3...\n')
-            system("pip3 install vtk==9.3.1"); }
-          
+            status2 = system(paste0(Sys.getenv('RETICULATE_PYTHON')," -m pip3 install vtk==9.3.1")); 
+          if (status2 != 0) {message("vtk 9.3.1 was not installed successfully. This may cause further issues.")}
+          }
           
           message('Please restart R after Python installation for its environment to be properly detected by reticulate.')
           
-        }
-        else { stop('VertexWiseR will not work properly without Miniconda or a suitable version of Python for reticulate.\n\n')}
-        
-      }} 
+      }
+      else { 
+        stop('VertexWiseR will not work properly without Miniconda or a suitable version of Python for reticulate.\n\n')
+      }
+      
+    } else 
+    {
+    reticulate::py_config() #if py_discover_config() successful, initialise for py_available to return TRUE next time
+    }} 
     
     #################################################################
     ##################################################################
@@ -239,7 +304,7 @@ VWRfirstrun=function(requirement="any", n_vert=0, promptless=FALSE)
     {
       missingobj=1
       
-      prompt = utils::menu(c("Yes", "No"), title="\n The Brainstat package could not be found in your Python/Conda environment. It is needed for vertex-wise linear models and the surface plotter to work. \n Do you want Brainstat (v0.4.2) to be installed now (~1.65 MB)? The NiMARE (~20.4 MB) and Brainspace (~84.2 MB) libraries and other BrainStat dependencies will automatically be installed within your Python library.")
+      prompt = utils::menu(c("Yes", "No"), title="\nThe Brainstat package could not be found in your Python/Conda environment. It is needed for vertex-wise linear models and the surface plotter to work. \n Do you want Brainstat (v0.4.2) to be installed now (~1.65 MB)? The NiMARE (~20.4 MB) and Brainspace (~84.2 MB) libraries and other BrainStat dependencies will automatically be installed within your Python library.")
       if (prompt==1)
       {	
         
@@ -263,8 +328,7 @@ VWRfirstrun=function(requirement="any", n_vert=0, promptless=FALSE)
       } 
       else {
         stop('VertexWiseR will not work properly without BrainStat.\n\n')}
-    } 
-    
+    }
     
     ###############################################################################################################################
     #check if BrainStat fsaverage/parcellation templates are installed (stops only if function needs it)
@@ -459,10 +523,7 @@ if (requirement!="python/conda only" & requirement!='conda/brainstat')
     
     if (promptless==TRUE) {if(!exists('missingobj'))
     { message('No system requirements are missing. \u2713 \n') }}
-    
-  } 
-  
-  
+  }
   
   #############################################################
   #############################################################
@@ -479,7 +540,7 @@ if (requirement!="python/conda only" & requirement!='conda/brainstat')
     non_interactive="Run VWRfirstrun() in an interactive R session to check for the missing system requirements and to install them."
     
     #miniconda or python missing?
-    if(is(tryCatch(reticulate::py_config(), error=function(e) e))[1] == 'simpleError') #fast, though less reliable check if Python is in a custom location. It will work if VWRfirstrun() was run once already
+    if(reticulate::py_available()==FALSE) #fast, but unreliable check if Python is in a custom location. It will work if VWRfirstrun() was run once already
     {
       if (is.null(reticulate::py_discover_config()) |
           is(tryCatch(reticulate::py_discover_config(), error=function(e) e))[1] == 'simpleError') #slow but reliable check if first time 
@@ -490,7 +551,10 @@ if (requirement!="python/conda only" & requirement!='conda/brainstat')
         { non_interactive=paste0(missingobj,non_interactive)
         return(non_interactive)
         } else {message(missingobj)}
-      }} 
+      } else {
+        reticulate::py_config #needed for py_available to return TRUE at the next check if py_discover_config was successful
+      }
+    } 
     
     #brainstat missing
     if (!reticulate::py_module_available("brainstat")