+++ R/vertwise.R
@@ -26,7 +26,7 @@
 #' @param surf_data A N x V matrix object containing the surface data (N row for each subject, V for each vertex), in fsaverage5 (20484 vertices), fsaverage6 (81924 vertices), fslr32k (64984 vertices) or hippocampal (14524 vertices) space. See also Hipvextract(), SURFvextract() or FSLRvextract output formats. Alternatively, a string object containing the path to the surface object (.rds file) outputted by extraction functions may be given.
 #' @param p A numeric object specifying the p-value to threshold the results (Default is 0.05)
 #' @param atlas A numeric integer object corresponding to the atlas of interest.  1=Desikan, 2=Destrieux-148, 3=Glasser-360, 4=Schaefer-100, 5=Schaefer-200, 6=Schaefer-400. Set to `1` by default. This argument is ignored for hippocampal surfaces.
-#' @param smooth_FWHM A numeric vector object specifying the desired smoothing width in mm 
+#' @param smooth_FWHM A numeric vector object specifying the desired smoothing width in mm. It should not be specified if the surf_data has been smoothed previously with smooth_surf(), because this result in surf_data being smoothed twice.
 #' @param VWR_check A boolean object specifying whether to check and validate system requirements. Default is TRUE.
 #'
 #' @returns A list object containing the cluster level results, unthresholded t-stat map, thresholded t-stat map, positive, negative and bidirectional cluster maps, and a FDR-corrected p-value map.
@@ -120,8 +120,14 @@ RFT_vertex_analysis=function(model,contrast, random, formula, formula_dataset, s
     } else 
     {stop("data vector should only contain 20484 (fsaverage5), 81924 (fsaverage6), 64984 (fslr32k) or 14524 (hippocampal vertices) columns")}
   
+  #Solves the "no visible binding for global variable" issue
+  . <- SLM <- NULL 
+  . <- FixedEffect <- NULL
+  . <- MixedEffect <- NULL
+  
   ##import python libaries
-  brainstat.stats=reticulate::import("brainstat.stats", delay_load = TRUE)
+  reticulate::source_python(paste0(system.file(package='VertexWiseR'),'/python/brainstat.stats.SLM_VWR.py'))
+  
 
   ##fitting model
   #preparing mask for model
@@ -142,14 +148,13 @@ RFT_vertex_analysis=function(model,contrast, random, formula, formula_dataset, s
   data_dir=paste0(brainstat_data_path,'/brainstat_data/surface_data/')
   
   #define model to fit
-  if(missing(random)) {model0=brainstat.stats$terms$FixedEffect(model, "_check_categorical" = FALSE)}
-  else {model0=brainstat.stats$terms$MixedEffect(ran = as.factor(random),fix = model,"_check_categorical" = FALSE)}
+  if(missing(random)) {model0=FixedEffect(model, "_check_categorical" = FALSE)}
+  else {model0=MixedEffect(ran = as.factor(random),fix = model,"_check_categorical" = FALSE)}
   
-  #Solves the "no visible binding for global variable" issue
-  . <- SLM <- NULL 
+
   #read version of SLM that allows to specify the directory for the
   #fetch_template_surface option
-  reticulate::source_python(paste0(system.file(package='VertexWiseR'),'/python/brainstat.stats.SLM_VWR.py'))
+  
   
   model=SLM(model = model0,
             contrast=contrast,